Separate app configurators in multiple files
